+++ CHANGELOG.md
@@ -16,6 +16,7 @@ History:
 
 ### v1.1.5 -> v1.x.x (2010/MM/DD)
 
+ - [!] KISSY.Event.on 第一个参数只支持单个原生 dom 节点或原生 dom 节点数组以及选择器字符串
  - [!] calendar 模块，S.use('calendar', fn) 需要写成 S.use('calendar+css', fn), 否则不会自动加载样式
  - [*] loader 中，对内置模块无需配置，采用约定，去除 mods.js
  - [*] S.clone 增加防循环引用